requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic-settings==2.1.0
elasticsearch==8.10.1
python-multipart==0.0.6
pypdf==3.17.4
pillow==10.1.0
cryptography==44.0.0
